Updated NuGet packages

+++ FFXIVTool/Views/CharacterDetailsView3.xaml.cs
@@ -1050,7 +1050,7 @@ namespace FFXIVTool.Views
             MemoryManager.Instance.MemLib.writeMemory(MemoryManager.GetAddressString(MemoryManager.Instance.CameraAddress, Settings.Instance.Character.FOV2), "float", "0");
             CharacterDetails.Max.value = 20;
             MemoryManager.Instance.MemLib.writeMemory(MemoryManager.GetAddressString(MemoryManager.Instance.CameraAddress, Settings.Instance.Character.Max), "float", "20");
-            CharacterDetails.Min.value = 1.50f;
+            CharacterDetails.Min.value = (float)1.50;
             MemoryManager.Instance.MemLib.writeMemory(MemoryManager.GetAddressString(MemoryManager.Instance.CameraAddress, Settings.Instance.Character.Min), "float", "1.50");
             CharacterDetails.CZoom.value = 20;
             MemoryManager.Instance.MemLib.writeMemory(MemoryManager.GetAddressString(MemoryManager.Instance.CameraAddress, Settings.Instance.Character.CZoom), "float", "20");
